% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r, R/input.R
\docType{data}
\name{input}
\alias{input}
\title{Plik zawirajacy dane zewnetrznych koniecznych do dzialania pakietu}
\format{
Szczegolowe opisy danych dstepne sa w ich dokumentacji

Ramka danych zawiera:
\describe{
\item{Nat}{Natezenie ruchu}
\item{Segment}{Rodzaj pojazdow}
\item{Fuel}{Typ paliwa}
\item{Technology}{Technologia silnika}
}
}
\source{
Obiekt wskazniki powstal na podstawie arkusza z danymi pochodzacymi
z EEA - Europejskiej Agnecji Srodowiska.
Obiekt input zostal przygotowany wczesniej i zapisany do obiektu
}
\usage{
input

input
}
\description{
Obiekt jest losowo wygenerowanym wektorem.
Pozostale dane sa wybrane z obiektu wskazniki.
}
\examples{
\dontrun{
 plikdane
}
\dontrun{
plikdane
}
}
\keyword{datasets}
