% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obl_emisji.R
\name{obl_emisji}
\alias{obl_emisji}
\title{Obliczanie emisji niskiej - transport}
\usage{
obl_emisji(
  dane = input,
  kategoria = "Passenger Cars",
  euro = "Euro 5",
  mode = "",
  substancja = c("CO", "EC")
)
}
\arguments{
\item{dane}{data.frame}

\item{kategoria}{hcar}

\item{euro}{char}

\item{mode}{vector}

\item{substancja}{vector}
}
\value{

}
\description{
Obliczanie emisji niskiej - transport
}
