% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiz_emisji.R
\name{wiz_emisji}
\alias{wiz_emisji}
\title{Funkcja do Wizualizacji danych dostarczonych z pakietem}
\usage{
wiz_emisji(
  df = wskazniki,
  kat1 = Category,
  kat2 = "Passenger Cars",
  kategoryczna = Segment,
  numeryczna = Gamma
)
}
\arguments{
\item{df}{data.frame}

\item{kat1}{char}

\item{kat2}{char}

\item{kategoryczna}{char}

\item{numeryczna}{char}
}
\description{
Funkcja do Wizualizacji danych dostarczonych z pakietem
}
\examples{
#Wygenerowanie wykresu dla danych domuslnych
wiz_emisji()

}
