% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{input}
\alias{input}
\title{Plik zawirajacy dane zewnetrznych koniecznych do dzialania pakietu}
\format{
data.frame
\describe{

}
}
\source{
Plik wskazniki pochodzi z EEA - Europejskiej Agenji Srodowiska
}
\usage{
input
}
\description{
Plik zawirajacy dane zewnetrznych koniecznych do dzialania pakietu
}
\examples{
\dontrun{
 plikdane
}
}
\keyword{datasets}
