% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{input}
\alias{input}
\title{Plik zawirajacy dane zewnetrznych koniecznych do dzialania pakietu}
\format{
data.frame
\describe{
\item{Category}{Kategoria samochodu}
\item{Fuel}{Paliwo}
\item{Segment}{Segment - wielkosc}
\item{Euro.Standard}{Standard EURO}
\item{Technology}{Technologia oczyszczania}
\item{Pollutant}{Rodzaj zanieczyszczenia}
\item{Mode}{Mode}
\item{Alpha}{Parametr Alpha}
\item{Beta}{Parametr Beta}
\item{Gamma}{Parametr Gamma}
\item{Delta}{Parametr Delta}
\item{Epsilon}{Parametr Epsilon}
\item{Zita}{Parametr Zita}
\item{Hta}{Parametr Hta}
\item{Reduction}{Redukcja}
\item{Bio}{Bioredukcja}
\item{Procent}{Wielkosc procentowa}
}
}
\source{
Plik wskazniki pochodzi z EEA - Europejskiej Agenji Srodowiska
}
\usage{
input
}
\description{
Plik zawirajacy dane zewnetrznych koniecznych do dzialania pakietu
}
\examples{
\dontrun{
 plikdane
}
}
\keyword{datasets}
