% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obl_emisji.R
\name{obl_emisji}
\alias{obl_emisji}
\title{Title}
\usage{
obl_emisji(
  dane = input,
  kategoria = "Passenger Cars",
  euro = "Euro 5",
  mode = "",
  substancja = c("CO", "EC")
)
}
\arguments{
\item{substancja}{}
}
\value{

}
\description{
Title
}
