% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obl_emisji.R
\name{obl_emisji}
\alias{obl_emisji}
\title{Funkcja do bliczanie emisji pochodzacej z transportu}
\usage{
obl_emisji(
  dane = input,
  kategoria = "Passenger Cars",
  euro = "Euro 5",
  mode = "",
  substancja = c("CO", "EC")
)
}
\arguments{
\item{dane}{data.frame - obiekt wejsciowy}

\item{kategoria}{char - kategoria pojazdu}

\item{euro}{char - norma spalin}

\item{mode}{char - tryb jazdy}

\item{substancja}{char - mitowane zanieczyszczenie}
}
\description{
Funkcja oblicza emisje na korzystajac z danych zawartych w obiekcie
\code{wskazniki}, a tazke z obiektu \code{input}.
W obiekcie \code{wskazniki} zapisane sa dane o emisji, prygotowane wczesniej.
W obiekcie \code{input} zapisane sa stworzone losowo dane o natezenia ruchu.
}
\details{
Wzor do obliczen emisji:
Emisja = Nat x ((Alpha  x Procent^2 + Beta x Procent + Gamma +
(Delta/Procent)) / (Epsilon x Procent^2 + Zita x Procent + Hta) x (1- Reduction))
}
\examples{
# Obliczenie emisji dla parametrow domyslnych
  tmp <- obl_emisji()

}
