% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiz_emisji.R
\name{wiz_emisji}
\alias{wiz_emisji}
\title{Wizualizacja danych dostarczonych z pakietem wg wskazanych parametrow}
\usage{
wiz_emisji(
  df = wskazniki,
  kat1 = Category,
  kat2 = "Passenger Cars",
  kategoryczna = Segment,
  numeryczna = Gamma
)
}
\arguments{
\item{df}{data.frame}

\item{kat1}{charakter}

\item{kat2}{charakter}

\item{kategoryczna}{charakter}

\item{numeryczna}{charakter
#'}
}
\description{
Wizualizacja danych dostarczonych z pakietem wg wskazanych parametrow
}
\examples{
p <- wyk(df = wskazniki,
         kat1 = Pollutant,
         kat2 = "CH4",
         kategoryczna = Category,
         numeryczna = Gamma)

}
